# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AllowListUser:
    properties:
      ignores_player_limit:
        type: boolean
      name:
        type: string
      xuid:
        type: string
    type: object
...
  domain.ServerInfo:
    properties:
      active:
        type: boolean
      allow_list: {}
      create_status: {}
      created_at:
        description: 创建时间
        type: string
      description:
        description: 服务器描述
        type: string
      exist:
        type: boolean
      id:
        description: 主键
        type: integer
      name:
        description: 服务器名称
        type: string
      server_properties: {}
      updated_at:
        description: 更新时间
        type: string
      version:
        description: 版本
        type: string
      world_name:
        description: 世界名称
        type: string
    type: object
...
  errors.AppError:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 返回数据
      message:
        description: 错误信息
        type: string
    type: object
...
  handler.ConsoleLogResponse:
    properties:
      content:
        type: string
    type: object
  handler.CreateServerRequest:
    properties:
      description:
        type: string
      name:
        type: string
      version:
        type: string
      world_name:
        type: string
    required:
    - name
    type: object
...
  handler.UpdateServerRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
host: localhost:11223
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: MC Dashboard 后端服务API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MC Dashboard API
  version: "1.0"
paths:
...
  /servers:
    get:
      consumes:
      - application/json
      description: 获取所有服务器的列表，支持分页和排序
      parameters:
      - default: 0
        description: 页码，从1开始
        in: query
        name: page
        type: integer
      - default: 0
        description: 每页数量
        in: query
        name: size
        type: integer
      - default: desc
        description: 排序方向，asc或desc
        in: query
        name: order
        type: string
      - default: created_at
        description: 排序字段，支持id/name/version
        in: query
        name: orderBy
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器列表
          schema:
            items:
              $ref: '#/definitions/domain.ServerInfo'
            type: array
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取服务器列表
      tags:
      - 服务器
    post:
      consumes:
      - application/json
      description: 创建一个新的Minecraft服务器
      parameters:
      - description: 服务器创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 服务器创建成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 创建服务器
      tags:
      - 服务器
  /servers/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的Minecraft服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器删除成功
          schema:
            type: string
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 删除服务器
      tags:
      - 服务器
    get:
      consumes:
      - application/json
      description: 根据服务器ID获取服务器详细信息
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器信息
          schema:
            $ref: '#/definitions/domain.ServerInfo'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取服务器信息
      tags:
      - 服务器
    put:
      consumes:
      - application/json
      description: 修改服务器的配置信息
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      - description: 服务器更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 服务器编辑成功
          schema:
            type: string
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 编辑服务器
      tags:
      - 服务器
  /servers/{id}/allowlist:
    get:
      consumes:
      - application/json
      description: 获取指定服务器的白名单用户列表
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 白名单用户列表
          schema:
            items:
              $ref: '#/definitions/domain.AllowListUser'
            type: array
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取白名单列表
      tags:
      - 服务器
  /servers/{id}/allowlist/{username}:
    delete:
      consumes:
      - application/json
      description: 从服务器的白名单中删除指定用户
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除白名单用户成功
          schema:
            type: string
        "404":
          description: 服务器不存在或用户不在白名单中
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 删除白名单用户
      tags:
      - 服务器
    post:
      consumes:
      - application/json
      description: 将指定用户添加到服务器的白名单中
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      - description: 用户名
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 添加白名单用户成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 添加白名单用户
      tags:
      - 服务器
  /servers/{id}/console_log:
    get:
      consumes:
      - application/json
      description: 获取指定服务器的控制台日志
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: 获取最后几行日志
        in: query
        name: line
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 日志内容
          schema:
            $ref: '#/definitions/handler.ConsoleLogResponse'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取控制台日志
      tags:
      - 服务器
  /servers/{id}/server_properties:
    get:
      consumes:
      - application/json
      description: 获取指定服务器的所有属性配置
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器属性
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 获取服务器属性
      tags:
      - 服务器
    put:
      consumes:
      - application/json
      description: 更新指定服务器的属性配置
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      - description: 服务器属性
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 服务器属性更新成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 更新服务器属性
      tags:
      - 服务器
  /servers/{id}/start:
    post:
      consumes:
      - application/json
      description: 启动指定的Minecraft服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器启动成功
          schema:
            type: string
        "400":
          description: 服务器已经在运行
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 启动服务器
      tags:
      - 服务器
  /servers/{id}/stop:
    post:
      consumes:
      - application/json
      description: 停止指定的Minecraft服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 服务器停止成功
          schema:
            type: string
        "400":
          description: 服务器未在运行
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: 服务器不存在
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/errors.AppError'
      security:
      - Bearer: []
      summary: 停止服务器
      tags:
      - 服务器
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 请输入Bearer token，格式为：Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"